build station list and return a weatherdata object

+++ CurrentWeatherApp/CurrentWeatherAPI/src/model/WeatherData/StationData.cs
@@ -3,9 +3,9 @@ namespace CurrentWeatherAPI.src.model.WeatherData
     public class StationData
     {
         public required string Name { get; set; }
-        public required long Latitude { get; set; }
-        public required long Altitude { get; set; }
-        public required long Longitude { get; set; }
+        public required double Latitude { get; set; }
+        public required double Altitude { get; set; }
+        public required double Longitude { get; set; }
         public required long TimeGathered { get; set; }
         public required string Temperature { get; set; }
         public required string AirQuality { get; set; }
